+++ gpu-video/examples/print_hw_capabilities.rs
@@ -11,19 +11,19 @@ fn main() {
 
     tracing::subscriber::set_global_default(subscriber).expect("Failed to initialize tracing");
 
-    let vulkan_instance = VideoInstance::new(&VideoInstanceDescriptor {
+    let video_instance = VideoInstance::new(&VideoInstanceDescriptor {
         enable_validations: true,
         ..Default::default()
     })
     .unwrap();
-    let vulkan_adapter = vulkan_instance
+    let video_adapter = video_instance
         .create_adapter(&VideoAdapterDescriptor::default())
         .unwrap();
-    let vulkan_device = vulkan_adapter
+    let video_device = video_adapter
         .create_device(&VideoDeviceDescriptor::default())
         .unwrap();
 
-    std::hint::black_box(vulkan_device);
+    std::hint::black_box(video_device);
 }
 
 #[cfg(not(vulkan))]
