+++ gpu-video/examples/transcode.rs
@@ -46,22 +46,22 @@ fn main() {
         ScalingAlgorithm::default()
     };
 
-    let instance = VideoInstance::new(&VideoInstanceDescriptor {
+    let video_instance = VideoInstance::new(&VideoInstanceDescriptor {
         enable_validations: true,
         ..Default::default()
     })
     .unwrap();
-    let adapter = instance
+    let video_adapter = video_instance
         .create_adapter(&VideoAdapterDescriptor::default())
         .unwrap();
-    let device = adapter
+    let video_device = video_adapter
         .create_device(&VideoDeviceDescriptor::default())
         .unwrap();
 
     let average_bitrate = 1_000_000;
     let max_bitrate = 1_200_000;
 
-    let params_h264 = device
+    let params_h264 = video_device
         .encoder_output_parameters_h264_high_quality(RateControl::VariableBitrate {
             average_bitrate,
             max_bitrate,
@@ -69,7 +69,7 @@ fn main() {
         })
         .unwrap();
 
-    let params_h265 = device
+    let params_h265 = video_device
         .encoder_output_parameters_h265_high_quality(RateControl::VariableBitrate {
             average_bitrate,
             max_bitrate,
@@ -77,7 +77,7 @@ fn main() {
         })
         .unwrap();
 
-    let mut transcoder = device
+    let mut transcoder = video_device
         .create_transcoder(TranscoderParameters {
             input_framerate: 30.into(),
             output_parameters: vec![
