+++ gpu-video/src/lib.rs
@@ -6,8 +6,8 @@ mod adapter;
 pub(crate) mod codec;
 #[cfg(vulkan)]
 mod device;
-#[cfg(vulkan)]
-#[cfg_attr(not(feature = "wgpu"), allow(dead_code))]
+// TODO: cfg(vulkan) will not be needed once we add proper metal support
+#[cfg(all(vulkan, feature = "wgpu"))]
 mod global_registry;
 #[cfg(vulkan)]
 mod instance;
